+++ scripts/script.js
@@ -1,7 +1,9 @@
 /**
  * Script runner.
  */
+const process = require('process');
 const console = require('console');
+const path = require('path');
 const exec = require('child_process').exec;
 
 const run = (command) => {
@@ -22,7 +24,10 @@ const run = (command) => {
   });
 };
 
+const srcDirPath = process.env.npm_package_config_src_dir_path;
+const testBootstrap = path.join(srcDirPath, 'js', '__test__', 'index.js');
+
 module.exports = {
   run,
-  mochaOpts: '--recursive --compilers js:babel-register --require ./src/js/__test__'
+  mochaOpts: `--recursive --compilers js:babel-register --require ${testBootstrap}`
 };